+++ packages/smarthr-ui/src/components/SegmentedControl/client/SegmentedControl.tsx
@@ -11,9 +11,9 @@ import {
 } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useCallbackRefCleanupForReact18 } from '../../hooks/client/useCallbackRefCleanupForReact18'
-import { useLatest } from '../../hooks/useLatest'
-import { Button } from '../Button'
+import { useCallbackRefCleanupForReact18 } from '../../../hooks/client/useCallbackRefCleanupForReact18'
+import { useLatest } from '../../../hooks/useLatest'
+import { Button } from '../../Button'
 
 export type Option = {
   /** 選択時に返される値 */
