+++ packages/smarthr-ui/src/components/SegmentedControl/client/index.ts
@@ -1,0 +1,1 @@
+export { SegmentedControl, type Option } from './SegmentedControl'
