+++ packages/smarthr-ui/src/components/SegmentedControl/index.ts
@@ -1,1 +1,1 @@
-export { SegmentedControl, type Option as SegmentedControlOption } from './SegmentedControl'
+export { SegmentedControl, type Option as SegmentedControlOption } from './client'
