+++ packages/smarthr-ui/src/components/SegmentedControl/stories/SegmentedControl.stories.tsx
@@ -8,7 +8,7 @@ import {
   FaTableIcon,
 } from '../../Icon'
 import { Stack } from '../../Layout'
-import { SegmentedControl } from '../SegmentedControl'
+import { SegmentedControl } from '../client'
 
 import type { Meta, StoryObj } from '@storybook/react-vite'
