+++ packages/smarthr-ui/src/components/SegmentedControl/stories/VRTSegmentedControl.stories.tsx
@@ -2,7 +2,7 @@ import { userEvent } from 'storybook/test'
 
 import { FaChartAreaIcon, FaChartBarIcon, FaChartLineIcon } from '../../Icon'
 import { Stack } from '../../Layout'
-import { type Option, SegmentedControl } from '../SegmentedControl'
+import { type Option, SegmentedControl } from '../client'
 
 import type { StoryObj } from '@storybook/react-vite'
